requirements.txt
pandas 
matplotlib 
seaborn 
python-dotenv 
requests
pandas
json